src/backend/pyproject.toml
[tool.poetry]
name = "bioaccess-backend"
version = "1.0.0"
description = "Sistema de Autenticação Biométrica Facial com RBAC"
authors = ["BioAccess Team"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.115.0"
uvicorn = {extras = ["standard"], version = "0.32.0"}
pydantic-settings = "2.6.0"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
sqlalchemy = "2.0.36"
psycopg2-binary = "2.9.10"
opencv-python = "4.10.0.84"
numpy = "2.1.3"
pillow = "11.0.0"
deepface = "0.0.93"
httpx = "0.28.0"
python-multipart = "0.0.9"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-asyncio = "0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
